# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: UseCase Manager
  description: AI-powered use case extraction from workshop transcripts
  version: 0.1.0
paths:
...
  /use-cases/{use_case_id}/comments:
    get:
      tags:
      - comments
      summary: List Comments
      description: List comments on a use case
      operationId: list_comments_use_cases__use_case_id__comments_get
      parameters:
      - name: use_case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Use Case Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Comments Use Cases  Use Case Id  Comments Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
    post:
      tags:
      - comments
      summary: Create Comment
      description: Add a comment to a use case
      operationId: create_comment_use_cases__use_case_id__comments_post
      parameters:
      - name: use_case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Use Case Id
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                body:
                  type: string
                  title: Body
              type: object
              required:
              - body
              title: CommentCreate
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                    format: uuid
                    title: Id
                  body:
                    type: string
                    title: Body
                  use_case_id:
                    type: string
                    format: uuid
                    title: Use Case Id
                  author_id:
                    type: string
                    format: uuid
                    title: Author Id
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - id
                - body
                - use_case_id
                - author_id
                - created_at
                - updated_at
                title: CommentResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
  /use-cases/{use_case_id}/comments/{comment_id}:
    patch:
      tags:
      - comments
      summary: Update Comment
      description: "Edit a comment (only own comment, or admin)"
      operationId: update_comment_use_cases__use_case_id__comments__comment_id__patch
      parameters:
      - name: use_case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Use Case Id
      - name: comment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Comment Id
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                body:
                  type: string
                  title: Body
              type: object
              required:
              - body
              title: CommentUpdate
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                    format: uuid
                    title: Id
                  body:
                    type: string
                    title: Body
                  use_case_id:
                    type: string
                    format: uuid
                    title: Use Case Id
                  author_id:
                    type: string
                    format: uuid
                    title: Author Id
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - id
                - body
                - use_case_id
                - author_id
                - created_at
                - updated_at
                title: CommentResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
    delete:
      tags:
      - comments
      summary: Delete Comment
      description: "Delete a comment (only own comment, or admin)"
      operationId: delete_comment_use_cases__use_case_id__comments__comment_id__delete
      parameters:
      - name: use_case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Use Case Id
      - name: comment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Comment Id
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      responses:
        "204":
          description: Successful Response
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
...
components:
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/jwt/login